fix: fix code review
