--- src/components/DragAndDrop/DragAndDrop.scss
@@ -1,30 +1,0 @@
-.list > div {
-  display: flex;
-  flex-direction: column;
-}
-
-.item {
-  width: 180px;
-  padding: 10px;
-  margin: 10px auto 10px 10px;
-  //background: lightseagreen;
-  color: white;
-  font-family: sans-serif;
-  border-radius: 4px;
-  display: inline-block;
-  position: relative;
-  box-shadow: 0 1px 0 rgba(9, 30, 66, 0.25);
-  cursor: grab;
-}
-
-.item:hover{
-  opacity: 0.7;
-}
-
-.flip-list-move {
-  transition: transform .2s;
-}
-
-.over {
-  opacity: .6;
-}
