--- src/components/DragAndDrop/index.js
@@ -1,4 +1,0 @@
-import { createApp } from 'vue';
-import Example from './Example.vue';
-
-createApp(Example).mount('#app');
